fix: go install needs main.go file
