+++ cmd/doc.go
@@ -1,0 +1,2 @@
+// Package cmd contains the gsemver CLI implementation
+package cmd
