--- cmd/gsemver/doc.go
@@ -1,2 +1,0 @@
-// Package gsemver contains the gsemver CLI implementation
-package gsemver
