--- cmd/gsemver/gsemver.go
@@ -1,7 +1,0 @@
-package gsemver
-
-// Run runs the command
-func Run() error {
-	cmd := newDefaultRootCommand()
-	return cmd.Execute()
-}
