+++ cmd/root.go
@@ -1,4 +1,4 @@
-package gsemver
+package cmd
 
 import (
 	"io"
